Only reset trips with changed locations.

+++ ui/app/controllers/locations_controller.rb
@@ -103,8 +103,8 @@ class LocationsController < ApplicationController
         "order by st_distance(trips.start_location, locations.geolocation) limit 1) from trips) as start_location_search " +
         "where trips.id = start_location_search.trip_id returning id, start_location_id")].map do |changed_trip|
         if current_trips[changed_trip['id']]['start_location_id'] != changed_trip['start_location_id']
-          puts "Trip #{trip['id']} start location changed from #{current_trips[changed_trip['id']]['start_location_id']}\
-            #{changed_trip['start_location_id']} to "
+          puts "Trip #{trip['id']} start location changed from #{current_trips[changed_trip['id']]['start_location_id']} \
+            to #{changed_trip['start_location_id']}"
           Trip.find(trip['id']).trip_detail.destroy
         end
       end
@@ -113,9 +113,12 @@ class LocationsController < ApplicationController
         "(select trips.id as trip_id, (select locations.id location_id from locations " +
         "where ST_DWITHIN(trips.end_location, locations.geolocation, 200) " +
         "order by st_distance(trips.end_location, locations.geolocation) limit 1) from trips) as end_location_search " +
-        "where trips.id = end_location_search.trip_id returning id")].map do |trip|
-        puts "Purging trip #{trip['id']}"
-        Trip.find(trip['id']).trip_detail.destroy
+        "where trips.id = end_location_search.trip_id returning id")].map do |changed_trip|
+        if current_trips[changed_trip['id']]['end_location_id'] != changed_trip['end_location_id']
+          puts "Trip #{trip['id']} end location changed from #{current_trips[changed_trip['id']]['end_location_id']} \
+            to #{changed_trip['end_location_id']}"
+          Trip.find(trip['id']).trip_detail.destroy
+        end
       end
     end
